+++ crates/objc2/src/runtime/method_encoding_iter.rs
@@ -179,10 +179,10 @@ mod tests {
                 (
                     EncodingBox::Struct(
                         "bitfield".into(),
-                        Some(vec![
+                        vec![
                             EncodingBox::BitField(64, None),
                             EncodingBox::BitField(1, None),
-                        ]),
+                        ],
                     ),
                     Some(32),
                 ),